+++ ModularPathFinding/Dijkstra/Dijkstra.cpp
@@ -1,0 +1,8 @@
+#include "Dijkstra.h"
+#include "../Node.h"
+
+std::vector<NODE> PathAlgorithms::Dijkstra::GeneratePath(glm::ivec2 _start, glm::ivec2 _goal, std::vector<std::vector<NODE*>> _map, bool _allowDiagonal)
+{
+	
+	return std::vector<NODE>();
+};